pthread_t is a struct on BSD, cast to integer for printing
Test Plan:
make
fast_tests

DiffCamp Revision: 147143
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao
Revert Plan:
Ok

+++ src/runtime/eval/debugger/cmd/cmd_thread.cpp
@@ -173,7 +173,7 @@ void CmdThread::debuggerInfo(InfoVec &info) {
   Add(info, "Host",       Process::GetHostName());
   Add(info, "Binary",     Process::GetAppName());
   Add(info, "Process ID", FormatNumber("%lld", Process::GetProcessId()));
-  Add(info, "Thread ID",  FormatNumber("0x%llx", Process::GetThreadId()));
+  Add(info, "Thread ID",  FormatNumber("0x%llx", (int64)Process::GetThreadId()));
 }
 
 bool CmdThread::onServer(DebuggerProxy *proxy) {
